refactor: reorganize validation schemas
